requirements.txt
msgpack==1.1.0
numpy==2.2.4
packaging==24.2
pyserial==3.5
python-can==4.5.0
typing_extensions==4.13.0
wrapt==1.17.2
